+++ src/bot/custom_interactions/modal/support_system/account_recovery_modal.ts
@@ -8,7 +8,7 @@ import { CustomEmbed } from '@root/bot/common/message';
 
 import { CustomInteraction, CustomInteractionAccessLevel, CustomInteractionRunContext } from '@root/bot/common/managers/custom_interactions_manager';
 
-import { createSupportTicketChannel } from '@root/bot/handlers/support_system_handler';
+import { SupportTicketId, createSupportTicketChannel, support_categories } from '@root/bot/handlers/support_system_handler';
 
 //------------------------------------------------------------//
 
@@ -37,11 +37,13 @@ export default new CustomInteraction({
         const new_discord_user_id = interaction.fields.getTextInputValue('new_discord');
         const recovery_reason = interaction.fields.getTextInputValue('recovery_reason');
 
-        const support_channel = await createSupportTicketChannel(interaction.guild, interaction.member, 'RECOVERY');
+        const support_ticket_category = support_categories.find(({ id }) => id === SupportTicketId.Recovery)!;
+
+        const support_channel = await createSupportTicketChannel(interaction.guild, interaction.member, support_ticket_category);
 
         await interaction.editReply({
             content: [
-                'You selected Account Recovery!',
+                `You selected ${support_ticket_category.name}.`,
                 `Go to ${support_channel} to continue.`,
             ].join('\n'),
         });
