+++ src/bot/custom_interactions/modal/support_system/other_questions_modal.ts
@@ -8,17 +8,17 @@ import { CustomEmbed } from '@root/bot/common/message';
 
 import { CustomInteraction, CustomInteractionAccessLevel, CustomInteractionRunContext } from '@root/bot/common/managers/custom_interactions_manager';
 
-import { createSupportTicketChannel } from '@root/bot/handlers/support_system_handler';
+import { SupportTicketId, createSupportTicketChannel, support_categories } from '@root/bot/handlers/support_system_handler';
 
 //------------------------------------------------------------//
 
-const bot_database_support_staff_role_id = `${process.env.BOT_SUPPORT_STAFF_DATABASE_ROLE_ID ?? ''}`;
-if (bot_database_support_staff_role_id.length < 1) throw new Error('Environment variable: BOT_SUPPORT_STAFF_DATABASE_ROLE_ID; is not set correctly.');
+const bot_other_support_staff_role_id = `${process.env.BOT_SUPPORT_STAFF_OTHER_ROLE_ID ?? ''}`;
+if (bot_other_support_staff_role_id.length < 1) throw new Error('Environment variable: BOT_SUPPORT_STAFF_OTHER_ROLE_ID; is not set correctly.');
 
 //------------------------------------------------------------//
 
 export default new CustomInteraction({
-    identifier: 'product_transfer_modal',
+    identifier: 'other_questions_modal',
     type: Discord.InteractionType.ModalSubmit,
     data: undefined,
     metadata: {
@@ -31,16 +31,15 @@ export default new CustomInteraction({
 
         await interaction.deferReply({ ephemeral: true });
 
-        const products = interaction.fields.getTextInputValue('products');
-        const discord_transfer_to = interaction.fields.getTextInputValue('discord_transfer_to');
-        const roblox_transfer_to = interaction.fields.getTextInputValue('roblox_transfer_to');
-        const transfer_reason = interaction.fields.getTextInputValue('transfer_reason');
+        const question = interaction.fields.getTextInputValue('question');
 
-        const support_channel = await createSupportTicketChannel(interaction.guild, interaction.member, 'TRANSFERS');
+        const support_ticket_category = support_categories.find(({ id }) => id === SupportTicketId.Other)!;
+
+        const support_channel = await createSupportTicketChannel(interaction.guild, interaction.member, support_ticket_category);
 
         await interaction.editReply({
             content: [
-                'You selected Account Recovery!',
+                `You selected ${support_ticket_category.name}.`,
                 `Go to ${support_channel} to continue.`,
             ].join('\n'),
         });
@@ -49,7 +48,7 @@ export default new CustomInteraction({
             content: [
                 `${interaction.member}, welcome to your support ticket,`,
                 '',
-                `Our <@&${bot_database_support_staff_role_id}> support staff are unscheduled volunteers, so please be patient.`,
+                `Our <@&${bot_other_support_staff_role_id}> support staff are unscheduled volunteers, so please be patient.`,
                 '',
                 'If you have an urgent issue, like someone making death threats;',
                 'please @mention one of our high-ranked staff members!',
@@ -61,17 +60,8 @@ export default new CustomInteraction({
                         name: 'Inertia Lighting | Support System',
                     },
                     description: [
-                        '**What product(s) do you want to transfer?**',
-                        `${products}`,
-                        '',
-                        '**Discord user ID that you\'re transferring to?**',
-                        `${discord_transfer_to}`,
-                        '',
-                        '**Roblox account that you\'re transferring to?**',
-                        `${roblox_transfer_to}`,
-                        '',
-                        '**Why are you transferring your product(s)?**',
-                        `${transfer_reason}`,
+                        '**What can we help you with?**',
+                        `${question}`,
                     ].join('\n'),
                 }),
             ],
